backend/requirements.txt
langchain
langgraph
click
fastapi
uvicorn
python-dotenv
elevenlabs
google-generativeai
langchain-google-genai
sqlmodel
pyaudio
SpeechRecognition
python-multipart
